requirements.txt
# opencode-self-improving dependencies
gradio[mcp]>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
huggingface_hub>=0.20.0